Reworked tests
