+++ UnitTest/VisitorTests/0Prettyprint/Unit.test
@@ -1,0 +1,47 @@
+﻿Test single unit
+////
+unit unit1 {
+    ms => value;
+}
+////
+unit unit1 {
+    ms => value;
+}
+//////
+Test 2 subunits
+////
+unit unit2 {
+    ms => value;
+    s => value * 1000;
+}
+////
+unit unit2 {
+    ms => value;
+    s => value * 1000;
+}
+//////
+Test 2 units
+////
+unit unit3 {
+    ms => value;
+}
+unit unit4 {
+    s => value * 1000;
+}
+////
+unit unit3 {
+    ms => value;
+}
+unit unit4 {
+    s => value * 1000;
+}
+//////
+Should succeed with only units
+////
+unit unit5 {
+    ms => value;
+}
+////
+unit unit5 {
+    ms => value;
+}
