+++ UnitTest/VisitorTests/1Typecheck/Unit.test
@@ -1,0 +1,60 @@
+﻿Test single unit
+////
+unit unit1 {
+    ms => value;
+}
+////
+true
+//////
+Test 2 subunits
+////
+unit unit2 {
+    ms => value;
+    s => value * 1000;
+}
+////
+true
+//////
+Test 2 units
+////
+unit unit3 {
+    ms => value;
+}
+unit unit4 {
+    s => value * 1000;
+}
+////
+true
+//////
+Should succeed with only units
+////
+unit unit5 {
+    ms => value;
+}
+////
+true
+//////
+Legal unitTyped function
+////
+unit Time {
+    ms => value;
+    s => value * 1000;
+}
+
+int testValue;
+Time testUnit;
+
+Time functionTest(int a, string b){
+    testValue++;
+    return 10s;
+}
+
+prog {
+    testValue = 0;
+}
+loop {
+   Time testUnit;
+   testUnit = functionTest(5, "10");
+}
+////
+true
