+++ UnitTest/VisitorTests/2Codegen/EscapedChars.test
@@ -1,0 +1,27 @@
+﻿escaped chars
+////
+prog {
+    string n = "\n";
+    char a = '\a';
+    string b = "\b";
+    string f = "\f";
+    string r = "\r";
+    string t = "\t";
+    string v = "\v";
+    string q = "\?";
+    string slash = "\\";
+}
+////
+void loop() {
+}
+void setup() {
+    String n = String("\n");
+    char a = '\a';
+    String b = String("\b");
+    String f = String("\f");
+    String r = String("\r");
+    String t = String("\t");
+    String v = String("\v");
+    String q = String("\?");
+    String slash = String("\\");
+}
