+++ UnitTest/VisitorTests/2Codegen/Exp.test
@@ -1,0 +1,195 @@
+﻿expr plus test
+////
+func funexpr(int a){
+    int b = 5;
+    b = b + a;
+    decimal d = 3.5;
+    decimal e = 6.5 + d;
+    d = d + a;
+    
+    string s = "lmao";
+    string t = "xD";
+    s = s+a;
+    t = t+s; 
+    t = t+d;
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Ffunexpr(int a) {
+    int b = 5;
+    b = b + a;
+    float d = 3.5;
+    float e = 6.5 + d;
+    d = d + a;
+    String s = String("lmao");
+    String t = String("xD");
+    s = s + String(a);
+    t = t + s;
+    t = t + String(d);
+}
+//////
+expr minus testt
+////
+func untyped(int a){
+    int b = 5;
+    b = b-a;
+    decimal c = 3.5;
+    decimal d = 6.5;
+    d = d-c;
+    c = c-b;
+    
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Funtyped(int a) {
+    int b = 5;
+    b = b - a;
+    float c = 3.5;
+    float d = 6.5;
+    d = d - c;
+    c = c - b;
+}
+//////
+expr div mult mod
+////
+func untypedlmao(){
+    int a = 3;
+    int b = 5;
+    decimal cd = 3.5;
+    b = b * a;
+    decimal d = 6.5;
+    d = d * cd;
+    cd = cd / b;
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Funtypedlmao() {
+    int a = 3;
+    int b = 5;
+    float cd = 3.5;
+    b = b * a;
+    float d = 6.5;
+    d = d * cd;
+    cd = cd / b;
+}
+//////
+tern expr
+////
+func untypedlmao(){
+    bool b = true;
+    int c = 5;
+    int d = 4;
+    d = b? c : d;
+    
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Funtypedlmao() {
+    bool b = true;
+    int c = 5;
+    int d = 4;
+    d = b ? c : d;
+}
+//////
+logical AND OR expr
+////
+func untypedlmao() {
+    bool b = true;
+    bool a = false;
+    bool c = true;
+    bool d = a||b&&a||c;
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Funtypedlmao() {
+    bool b = true;
+    bool a = false;
+    bool c = true;
+    bool d = a || b && a || c;
+}
+//////
+logical compare expr
+////
+func untypedlmao(int a){
+    a = 5;
+    int b = 2;
+    int c = 7;
+    bool d = a == b || b < 7 ;
+    bool e = a != b && c > 5;
+    d = a >= 2 || b >= 3;
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Funtypedlmao(int a) {
+    a = 5;
+    int b = 2;
+    int c = 7;
+    bool d = a == b || b < 7;
+    bool e = a != b && c > 5;
+    d = a >= 2 || b >= 3;
+}
+//////
+suff-postfix expr 
+////
+func untypedlmao(){
+    int a = 5;
+    int b = 2;
+    decimal d = 3.5;
+    int c;
+    c = ++b + a;
+    b = b++ + a;
+    a = -a + b;
+    
+    bool rigtige = true;
+    bool falsk = !rigtige;  
+}
+func untypedlomo(int a){
+    a = 5;
+    int b = 2;
+    decimal d = 3.5;
+    int c;
+    c = --b + a;
+    b = b-- + a;
+}
+////
+void setup() {
+}
+void loop() {
+}
+void Funtypedlmao() {
+    int a = 5;
+    int b = 2;
+    float d = 3.5;
+    int c;
+    c = ++b + a;
+    b = b++ + a;
+    a = -a + b;
+    bool rigtige = true;
+    bool falsk = !rigtige;
+}
+void Funtypedlomo(int a) {
+    a = 5;
+    int b = 2;
+    float d = 3.5;
+    int c;
+    c = --b + a;
+    b = b-- + a;
+}
