+++ UnitTest/VisitorTests/2Codegen/FuncMin.test
@@ -1,0 +1,35 @@
+﻿????
+////
+prog {
+}
+loop {
+}
+string function3(char a, char b, char c) {
+    string d = (string)a;
+    return d;
+}
+decimal function4(int a) {
+    decimal d = 3.5 + a;
+    char g = 'o';
+    char b = 'b';
+    char c = 'c';
+    string test = function3(g, b, c);
+    return d;
+}
+////
+void setup() {
+}
+void loop() {
+}
+String Ffunction3(char a, char b, char c) {
+    String d = String(a);
+    return d;
+}
+float Ffunction4(int a) {
+    float d = 3.5 + a;
+    char g = 'o';
+    char b = 'b';
+    char c = 'c';
+    String test = Ffunction3(g, b, c);
+    return d;
+}
