+++ UnitTest/VisitorTests/2Codegen/GlobalScope.test
@@ -1,0 +1,11 @@
+﻿Global Declaration
+////
+int a = 7;
+int b;
+////
+void setup() {
+}
+void loop() {
+}
+int a = 7;
+int b;
