+++ UnitTest/VisitorTests/2Codegen/Unit.test
@@ -4,10 +4,6 @@ unit unit1 {
     ms => value;
 }
 ////
-unit unit1 {
-    ms => value;
-}
-////
 void setup() {
 }
 void loop() {
@@ -15,15 +11,8 @@ void loop() {
 float Uunit1ms(float value) {
     return value;
 }
-////
-true
 //////
 Test 2 subunits
-////
-unit unit2 {
-    ms => value;
-    s => value * 1000;
-}
 ////
 unit unit2 {
     ms => value;
@@ -40,17 +29,8 @@ float Uunit2ms(float value) {
 float Uunit2s(float value) {
     return value * 1000;
 }
-////
-true
 //////
 Test 2 units
-////
-unit unit3 {
-    ms => value;
-}
-unit unit4 {
-    s => value * 1000;
-}
 ////
 unit unit3 {
     ms => value;
@@ -69,14 +49,8 @@ float Uunit3ms(float value) {
 float Uunit4s(float value) {
     return value * 1000;
 }
-////
-true
 //////
 Should succeed with only units
-////
-unit unit5 {
-    ms => value;
-}
 ////
 unit unit5 {
     ms => value;
@@ -89,33 +63,3 @@ void loop() {
 float Uunit5ms(float value) {
     return value;
 }
-////
-true
-//////
-Legal unitTyped function
-////
-unit Time {
-    ms => value;
-    s => value * 1000;
-}
-
-int testValue;
-Time testUnit;
-
-Time functionTest(int a, string b){
-    testValue++;
-    return 10s;
-}
-
-prog {
-    testValue = 0;
-}
-loop {
-   Time testUnit;
-   testUnit = functionTest(5, "10");
-}
-
-////
-////
-////
-true
